+++ ADC_ex.c
@@ -2,13 +2,14 @@
 ** \author SMFSW
 ** \copyright MIT (c) 2017-2019, SMFSW
 ** \brief Simple extension for ADCs
+** \note Define USE_ADC_EX symbol at project level to use ADC_ex functionalities
 ** \note Configuration done calling ADC_cfg.h file
 **/
 /****************************************************************/
 #include "sarmfsw.h"
 #include "ADC_ex.h"
 
-#if defined(HAL_ADC_MODULE_ENABLED)
+#if defined(HAL_ADC_MODULE_ENABLED) && defined(USE_ADC_EX)
 /****************************************************************/
 
 
@@ -327,6 +328,6 @@ void HAL_ADC_ConvCpltCallback(ADC_HandleTypeDef * hadc)
 }
 
 
-/****************************************************************/
-#endif	/* defined(HAL_ADC_MODULE_ENABLED) */
-/****************************************************************/
+/******************************************************************/
+#endif	/* defined(HAL_ADC_MODULE_ENABLED) && defined(USE_ADC_EX) */
+/******************************************************************/